requirements.txt
# Core dependencies
torch>=2.0.0
diffusers>=0.25.0
transformers>=4.36.0
accelerate>=0.25.0
datasets>=2.14.0
wandb>=0.16.0

# Image processing
Pillow>=10.0.0
rasterio>=1.3.0
numpy>=1.24.0

# Development tools
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
pytest>=7.0.0

# Optional but recommended
tqdm>=4.65.0
matplotlib>=3.7.0
seaborn>=0.12.0 